Preserve circulars search params when returning from new circular form
Fixes #1105.
