+++ app/routes/circulars/index.tsx
@@ -24,10 +24,10 @@ import {
 import classNames from 'classnames'
 import { useState } from 'react'
 
-import { circularRedirect, put, search } from './circulars.server'
+import type { action } from '../circulars'
+import { circularRedirect, search } from './circulars.server'
 import Hint from '~/components/Hint'
 import { usePagination } from '~/lib/pagination'
-import { getFormDataString } from '~/lib/utils'
 
 import searchImg from 'app/theme/img/usa-icons-bg/search--white.svg'
 
@@ -53,15 +53,6 @@ export async function loader({ request: { url } }: DataFunctionArgs) {
   return { page, ...results }
 }
 
-export async function action({ request }: DataFunctionArgs) {
-  const data = await request.formData()
-  const body = getFormDataString(data, 'body')
-  const subject = getFormDataString(data, 'subject')
-  if (!body || !subject)
-    throw new Response('Body and subject are required', { status: 400 })
-  return await put(subject, body, request)
-}
-
 function getPageLink({
   page,
   query,
@@ -79,8 +70,8 @@ function getPageLink({
   if (startDate) searchParams.set('startDate', startDate)
   if (endDate) searchParams.set('endDate', endDate)
 
-  const searchParamsString = searchParams.toString()
-  return searchParamsString && `?${searchParamsString}`
+  const searchString = searchParams.toString()
+  return searchString && `?${searchString}`
 }
 
 function Pagination({
@@ -190,8 +181,8 @@ export default function () {
   const startDate = searchParams.get('startDate') ?? undefined
   const endDate = searchParams.get('endDate') ?? undefined
 
-  let searchParamsString = searchParams.toString()
-  if (searchParamsString) searchParamsString = `?${searchParamsString}`
+  let searchString = searchParams.toString()
+  if (searchString) searchString = `?${searchString}`
 
   const [inputQuery, setInputQuery] = useState(query)
   const clean = inputQuery === query
@@ -241,7 +232,7 @@ export default function () {
             />
           </Button>
         </Form>
-        <Link to="/circulars/new">
+        <Link to={`/circulars/new${searchString}`}>
           <Button
             type="button"
             className="height-4 padding-top-0 padding-bottom-0"
@@ -266,7 +257,7 @@ export default function () {
           <ol>
             {allItems.map(({ circularId, subject }) => (
               <li key={circularId} value={circularId}>
-                <Link to={`/circulars/${circularId}${searchParamsString}`}>
+                <Link to={`/circulars/${circularId}${searchString}`}>
                   {subject}
                 </Link>
               </li>
